requirements_phase1.txt
python-obd>=0.7.1
pyserial>=3.5
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0
